+++ CHANGELOG.md
@@ -1,5 +1,35 @@
 # Changelog
 All changes in this project will be noted in this file.
+
+## Version 3.0.0 [2020-05-21]
+> This release introduces breaking changes!
+
+- This release splits timer objects into two kinds: `SimpleTimer` and `ComplexTimer`.
+    `SimpleTimers` keep things simple: use `.start()` and `.stop()` to benchmark
+different operations, however you have to use the results of one benchmark
+immediately. This is ideal when you're benchmarking an operation and then printing
+or storing the results right away. However, if you want a timername/timer
+way of benchmarking, then `ComplexTimer` is for you. 
+- Also, the `run_through()`
+feature has now been deprecated and instead, you can use the function `run_benchmark()`
+which does the same thing. The function was moved out since it didn't rely on a 
+specific instance of a `SimpleTimer` (or previously `DevTime`) object.
+
+### Upgrading existing code
+- For any code that used `DevTime::new()` in versions prior to `3.0.0`, simply change it to `DevTime::new_simple()`.
+- For code that used the `run_through` feature introduced in `2.0.0` can continue to do so, but I recommend 
+upgrading to the `run_benchmark` function like shown below:
+```rust
+use devtimer::run_benchmark;
+fn main() {
+    let result_of_sleep = run_benchmark(10, || {
+        // Fake a slow operation
+        std::thread::sleep(std::time::Duration::from_secs(1));
+    });
+    result_of_sleep.print_stats();
+}
+```
+
 ## Version 2.0.0 [2020-05-09]
 > No breaking changes in this release